NCNN_Compression/requirements.txt
pyyaml>=6.0.1
numpy>=1.24.0
ultralytics>=8.1.0
torch-pruning>=1.4.0
onnx>=1.15.0
onnxruntime>=1.17.0
pillow>=10.0.0
onnxsim
torch